moved utility functions in util module
